requirements.txt
supabase
fastapi
uvicorn[standard]
python-dotenv